refactor: subject from spring oauth2 now give the user id

+++ src/main/java/com/anthony/calorie_counter/enums/UserRole.java
@@ -8,12 +8,12 @@ import java.util.List;
 
 @Getter
 public enum UserRole {
-    ROLE_USER(1),
-    ROLE_ADMIN(2);
+    ROLE_USER(1L),
+    ROLE_ADMIN(2L);
 
-    private final int role;
+    private final Long role;
 
-    UserRole(int role) { this.role = role; }
+    UserRole(Long role) { this.role = role; }
 
     public static UserRole getRoleFrom(String roleString) {
         try {
